+++ inngest-spring-boot-demo/src/test/java/com/inngest/springbootdemo/IntegrationTest.java
@@ -1,7 +1,9 @@
 package com.inngest.springbootdemo;
 
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -9,5 +11,7 @@ import java.lang.annotation.RetentionPolicy;
 @Retention(RetentionPolicy.RUNTIME)
 @EnabledIfSystemProperty(named = "test-group", matches = "integration-test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@Import(DemoTestConfiguration.class)
+@AutoConfigureMockMvc
 public @interface IntegrationTest {
 }
