+++ inngest-spring-boot-demo/src/test/java/com/inngest/springbootdemo/SendEventFunctionIntegrationTest.java
@@ -1,0 +1,47 @@
+package com.inngest.springbootdemo;
+
+import com.inngest.CommHandler;
+import com.inngest.Inngest;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.parallel.Execution;
+import org.junit.jupiter.api.parallel.ExecutionMode;
+import org.springframework.beans.factory.annotation.Autowired;
+
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@IntegrationTest
+@Execution(ExecutionMode.CONCURRENT)
+class SendEventFunctionIntegrationTest {
+    @BeforeAll
+    static void setup(@Autowired CommHandler handler) {
+        handler.register();
+    }
+
+    @Autowired
+    private DevServerComponent devServer;
+
+    static int sleepTime = 5000;
+
+    @Autowired
+    private Inngest client;
+
+    @Test
+    void testSendEventFunctionSendsEventSuccessfully() throws Exception {
+        String eventId = InngestFunctionTestHelpers.sendEvent(client, "test/send").first();
+
+        Thread.sleep(sleepTime);
+
+        RunEntry<Object> run = devServer.runsByEvent(eventId).first();
+
+        // Generic nested structures are frustrating to deserialize properly.
+        LinkedHashMap<String, ArrayList<String>> output = (LinkedHashMap<String, ArrayList<String>>) run.getOutput();
+        ArrayList<String> ids = output.get("ids");
+
+        assertEquals(run.getStatus(), "Completed");
+        assert !ids.isEmpty();
+    }
+}
